feat: removed setting RPC endpoint altogherher and altered the routes accordingly
