+++ src/pages/top-up/BankCardTopUpIndex.tsx
@@ -1,11 +1,11 @@
 import { Typography } from '@material-ui/core'
 import { ReactElement } from 'react'
-import Index from '.'
+import Balance from './Balance'
 import { ROUTES } from '../../routes'
 
 export function BankCardTopUpIndex(): ReactElement {
   return (
-    <Index
+    <Balance
       header={'Top-up with bank card'}
       title={'Use a bank card to buy xDAI to the funding wallet address below'}
       p={
